api/requirements.txt
####################################################
## Django required packages
####################################################

# Base project
django

# REST API for Django
djangorestframework

# Library for writing reusable URL query filters
django-filter

# Django middleware which permits cross-domain API requests
django-cors-headers

# WSGI HTTP server (useful for development, or if you don't plan on using any async features)
gunicorn



####################################################
## Django authentication (you can exclude any packages you don't plan on using)
####################################################

# Django LDAP authentication (need to download wheel to use on Windows)
django-auth-ldap

# Django PAM authentication
django-pam

# Django social authentication
social-auth-core
social-auth-app-django

# JWT authentication for DRF
djangorestframework-simplejwt



####################################################
## Required by Django or DRF
####################################################

# Postgres backend for Django database
psycopg2-binary

# Library with helpful timezone functions
pytz

# Regex support
regex

# Parses environment variables
environs

# Required for DRF
pynacl
pyyaml
uritemplate

# Required by some other packages
cryptography



####################################################
## Async Support
####################################################

# ASGI HTTP server with focus on speed
uvicorn[standard]

# Async library that has async version of various sync functions
asyncio

# Async HTTP library for making requests in an async environment
aiohttp



####################################################
## Optional requirements (will have to modify code to remove some references in some cases)
####################################################

# Provides colored output to logging
coloredlogs

# Provides managable arrays in the Django admin panel
django-better-admin-arrayfield

# Provides optional mixins and routers for DRF
drf-extensions

# Provides a validated hexadecimal color field
django-colorfield

# Library to make HTTP requests (sync only)
requests

